+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 ### Fixes
 
-[#2549: Fix header with product name focus and hover state length](https://github.com/alphagov/govuk-frontend/pull/2549)
-[#2573: Reinstate support for 'inherit' as a text colour](https://github.com/alphagov/govuk-frontend/pull/2573)
+- [#2549: Fix header with product name focus and hover state length](https://github.com/alphagov/govuk-frontend/pull/2549)
+- [#2573: Reinstate support for 'inherit' as a text colour](https://github.com/alphagov/govuk-frontend/pull/2573)
 
 ## 4.0.1 (Fix release)